backend/requirements.txt
asgiref==3.9.0
Django==5.2.4
django-cors-headers==4.7.0
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-filter==25.1
djangorestframework==3.16.0
dj-rest-auth==6.0.0
Pillow==10.1.0
psycopg2-binary==2.9.9
python-decouple==3.8
qrcode==7.4.2
sqlparse==0.5.3
typing_extensions==4.14.1
